fix(trusty-console): apply write-origin guard router-wide + bind-aware self-origin allowlist
Closes #3268
Closes #3269
